worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "numpy>=2.4.4",
    "scikit-learn>=1.8.0",
]
